# simd magnifyBy2 faster still