before:
`magnifyBy2 ........................ 10.087 ms     10.205 ms    ±0.115   x488`

after:
`magnifyBy2 ......................... 5.364 ms      5.585 ms    ±0.198   x888`